+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project is expected to follow [Semantic Versioning](https://semver.org/
 
 ### [Changed]
 
+- Normalize TAC instructions, eliminate variants of the same instruction.
 - Replace read/write builtins with explicit intrinsic calls in TAC.
 - Lower TAC variable access into explicit loads and stores.
 - Properly split flat TAC into well defined basic blocks.